bet_server_project/requirements.txt
Django==1.8.7
djangorestframework==3.3.1
